requirements.txt
langchain==0.3.26
langchain-community==0.3.26
langchain-text-splitters
langchain-openai==0.3.24
sentence-transformers==4.1.0
flask==3.1.1
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
hf_xet
tf-keras
-e .